+++ pymongo/asynchronous/mongo_client.py
@@ -34,6 +34,7 @@ from __future__ import annotations
 
 import contextlib
 import os
+import warnings
 import weakref
 from collections import defaultdict
 from typing import (
@@ -871,6 +872,7 @@ class AsyncMongoClient(common.BaseObject, Generic[_DocumentType]):
         )
 
         self._opened = False
+        self._closed = False
         self._init_background()
 
         if _IS_SYNC and connect:
@@ -1179,6 +1181,22 @@ class AsyncMongoClient(common.BaseObject, Generic[_DocumentType]):
         :param name: the name of the database to get
         """
         return database.AsyncDatabase(self, name)
+
+    def __del__(self) -> None:
+        """Check that this AsyncMongoClient has been closed and issue a warning if not."""
+        try:
+            if not self._closed:
+                warnings.warn(
+                    (
+                        f"Unclosed {type(self).__name__} opened at:\n{self._topology_settings._stack}"
+                        f"Call {type(self).__name__}.close() to safely shut down your client and free up resources."
+                    ),
+                    ResourceWarning,
+                    stacklevel=2,
+                    source=self,
+                )
+        except AttributeError:
+            pass
 
     def _close_cursor_soon(
         self,
@@ -1547,6 +1565,7 @@ class AsyncMongoClient(common.BaseObject, Generic[_DocumentType]):
         if self._encrypter:
             # TODO: PYTHON-1921 Encrypted MongoClients cannot be re-opened.
             await self._encrypter.close()
+        self._closed = True
 
     if not _IS_SYNC:
         # Add support for contextlib.aclosing.